requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
requests
python-dotenv
cryptography
Flask-Login
Werkzeug